move ironic capabilities tests to their own file
